Fixed: the id of the test should not be in the test (fix #1)
The save cleared everything because I moved the id of the test in the
test. On save, the id was reloaded from the test, so was still "new" and
on moving to a "new" test everything is cleared. The save worked on the
server side but content was wiped on completion on the client side.